Trait vessels::reflect::Reflected

pub trait Reflected: 'static { }

A trait object that has reflection data generated by #[object]

This trait should not be implemented manually by any third-party crate. Trait objects of traits annotated with #[object] will be marked with this trait to indicate their usability as reflected trait objects. Moreover, any type that implements Trait<dyn SomeTrait> where dyn SomeTrait: Reflected will have a generated implementation allowing it to satisfy SomeTrait.