Trait vessels::Kind

pub trait Kind: Any + Sized + Sync + Send + Unpin + 'static {
    type ConstructItem: Serialize + DeserializeOwned + Send + Sync + Unpin + 'static;
    type ConstructError: ErrorBound;
    type ConstructFuture: Future<Output = ConstructResult<Self>> + Sync + Send + 'static;
    type DeconstructItem: Serialize + DeserializeOwned + Send + Sync + Unpin + 'static;
    type DeconstructError: ErrorBound;
    type DeconstructFuture: Future<Output = DeconstructResult<Self>> + Sync + Send + 'static;
    fn construct<C: Channel<Self::ConstructItem, Self::DeconstructItem>>(
        channel: C
    ) -> Self::ConstructFuture;

    fn deconstruct<C: Channel<Self::DeconstructItem, Self::ConstructItem>>(
        self, 
        channel: C
    ) -> Self::DeconstructFuture;
}

A type with a bijection to over-the-wire data.

Kind is an advanced distributed object or RPC system that permits the over-the-wire serialization and deserialization of an implicitly flattened version of the complex nested structures required to produce a full type-level isomorphic representation of arbitrary composed data types.

Vessels provides Kind implementations for many primitive types from the standard library as well as futures, streams, a variety of boxed function types, and more.

Vessels also provides a derive macro that automatically generates Kind implementations for structs and enums in addition to the object macro for generating Kind implementations for trait objects of user-defined traits.

Authors of third-party crates are encouraged to derive or implement Kind or Kind providers for types their crates expose that might be useful over some form of wire boundary, be it network, IPC, or any other similar transport.

Associated Types

type ConstructItem: Serialize + DeserializeOwned + Send + Sync + Unpin + 'static

The item transmitted over the network to the construction task from deconstruction.

type ConstructError: ErrorBound

The failure condition of constructing a concrete type from communicated data.

type ConstructFuture: Future<Output = ConstructResult<Self>> + Sync + Send + 'static

The concrete future type returned by the construction process.

type DeconstructItem: Serialize + DeserializeOwned + Send + Sync + Unpin + 'static

The item transmitted over the network from the construction task to deconstruction.

type DeconstructError: ErrorBound

The failure condition of constructing a concrete type from communicated data.

type DeconstructFuture: Future<Output = DeconstructResult<Self>> + Sync + Send + 'static

The concrete future type returned by the deconstruction process. This is used to only to communicate failure of deconstruction and does not return a value.

Loading content...
Required methods

fn construct<C: Channel<Self::ConstructItem, Self::DeconstructItem>>(
channel: C
) -> Self::ConstructFuture

Constructs the Kind from the provided channel. This method should return immediately and, if necessary, move channel into some shim structure, async block, or other owned state specified by ConstructFuture.

fn deconstruct<C: Channel<Self::DeconstructItem, Self::ConstructItem>>(
self,
channel: C
) -> Self::DeconstructFuture

Moves out of the Kind and deconstructs on to the provided channel. As with construct, this method should return immediately.

Loading content...